+++ packages/jest/index.js
@@ -1,3 +1,6 @@
-const webConfig = require('./web/config')
+import {nativeConfig} from './native/config.js'
+import {webConfig} from './web/config.js'
 
-module.exports = {webConfig}
+export default {
+  projects: [webConfig, nativeConfig]
+}