+++ packages/jest/web/index.js
@@ -1,0 +1,9 @@
+const axe = require('./axe')
+const rtl = require('./rtl')
+const userEvent = require('./userEvent')
+
+module.exports = {
+  ...rtl,
+  ...axe,
+  userEvent
+}