+++ packages/primitives/src/helpers/props.ts
@@ -1,7 +1,7 @@
-import {InterUnitInternals} from '@interunit/config'
+import {getEnvironmentName} from '@interunit/config'
 import {platformStyleTranslation} from '@interunit/crossplatform'
 
-const ENVIRONMENT = InterUnitInternals.InterUnitInternalConfig.ENVIRONMENT.NAME
+const ENVIRONMENT = getEnvironmentName()
 
 const nativeOnlyProps = [
   'onLayout',
@@ -11,7 +11,8 @@ const nativeOnlyProps = [
   'accessibilityHint',
   'accessibilityRole',
   'accessibilityState',
-  'onPress'
+  'onPress',
+  'testID'
 ]
 
 const webOnlyProps = ['onClick']