+++ packages/primitives/src/primitive/Text/Text.tsx
@@ -1,3 +1,4 @@
+import {getEnvironmentName} from '@interunit/config'
 import {type MergeWithOverride} from '@interunit/toolbox'
 import React from 'react'
 import type {TextProps} from 'react-native'
@@ -42,6 +43,12 @@ const Text = React.forwardRef(
     }
 
     const filteredProps = filterPropsByEnvironment({props})
+
+    // Might make sense to move this along with other things to some
+    // sort of HOC pattern for matching parity on props across platforms
+    if (filteredProps.hidden && getEnvironmentName() === 'native') {
+      return
+    }
 
     return (
       <Text ref={forwardedRef} {...filteredProps}>