+++ packages/toolbox/src/hooks/useControlledState/useControlledState.ts
@@ -1,0 +1,33 @@
+import React from 'react'
+
+export type ControlledStateParams<V> = {
+  value: V
+  defaultValue?: never
+  onValueChange: (value: V) => void
+}
+
+export type UncontrolledStateParams<V> = {
+  value?: never
+  defaultValue: V
+  onValueChange?: never
+}
+
+export type UseControlledStateParams<V> =
+  | ControlledStateParams<V>
+  | UncontrolledStateParams<V>
+
+// TODO: Catch runtime error if props are invalid
+
+function useControlledState<V>(params: UseControlledStateParams<V>) {
+  const isStateAlreadyControlled = params.value !== undefined
+
+  if (isStateAlreadyControlled) {
+    return [params.value, params.onValueChange] as const
+  }
+
+  const [value, setValue] = React.useState<V>(params.defaultValue as V)
+
+  return [value, setValue] as const
+}
+
+export {useControlledState}