+++ packages/crossplatform/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @interunit/crossplatform
 
+## 0.0.5
+
+### Patch Changes
+
+- Tabs keyboard navigation improvements + overhaul to testing stack
+- Updated dependencies
+  - @interunit/toolbox@0.0.5
+  - @interunit/config@0.0.6
+
 ## 0.0.4
 
 ### Patch Changes