+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @interunit/tsconfig
 
+## 0.0.4
+
+### Patch Changes
+
+- Tabs keyboard navigation improvements + overhaul to testing stack
+
 ## 0.0.3
 
 ### Patch Changes